arm.h (CONDITIONAL_REGISTER_USAGE): Avoid signed/unsigned warning.
	* arm.h (CONDITIONAL_REGISTER_USAGE): Avoid signed/unsigned
	warning.
	* emit-rtl.c (gen_rtx_REG, set_mem_attributes_minus_bitpos,
	init_emit_once): Likewise.
	* flow.c (mark_regs_live_at_end, calculate_global_regs_live):
	Likewise.
	* function.c (assign_stack_temp_for_type): Likewise.
	* loop.c (loop_invariant_p): Likewise.
	* recog.c (push_operand): Likewise.
	* regclass.c (init_reg_sets_1): Likewise.
	* reload.c (update_auto_inc_notes): Likewise.
	* reload1.c (reload_as_needed, emit_input_reload_insns): Likewise.
	* stmt.c (expand_asm_operands): Likewise.
	* stor-layout.c (start_record_layout): Likewise.

cp:
	* class.c (layout_virtual_bases): Avoid signed/unsigned warning.

java:
	* jcf-write.c (generate_bytecode_insns): Avoid signed/unsigned
	warning.

From-SVN: r61389

+++ gcc/function.c
@@ -733,7 +733,7 @@ assign_stack_temp_for_type (mode, size, keep, type)
 	abort ();
       p->slot = assign_stack_local (mode,
 				    (mode == BLKmode
-				     ? CEIL_ROUND (size, align / BITS_PER_UNIT)
+				     ? CEIL_ROUND (size, (int) align / BITS_PER_UNIT)
 				     : size),
 				    align);
